refactor(opt): Phase 4.5-A — revert routine sources to e_str; stash handwritten e_fn as golden fixtures
Reverts the routine files migrated in Phases 4.2-4.4 (dcopf, dopf,
rted, ed, uc) back to the e_str-form authoring API. Per the design
discussion, the goal is opt-layer simplification not e_str
elimination — the codegen plan introduces an "ams prep" step that
generates pycode from the existing e_str DSL, restoring the terse
authoring while moving regex/eval out of the runtime path.

The handwritten e_fn callables produced during 4.2-4.4 are preserved
under tests/codegen/golden/ as the parity oracle for the codegen
work in 4.5-B onward. They captured every routine path (linear,
sparse @ Variable, multi-period broadcast, big-M, MinDur, mixin
composition) and are the right thing to validate the generator
against.

Kept on the opt-layer side (these stand independently of e_str↔e_fn):
- _EFormDescriptor mutex (e_str/e_fn cross-clear on assignment)
- OptzBase.e fall-through for e_fn form
- Objective.add_term composition helper
- documenter._tex_pre placeholder for missing e_str
- ModelParam, RoutineNS scaffolding, DPP diagnostic

340 tests still pass on the e_str path.
