+++ ams/routines/rted.py
@@ -5,8 +5,6 @@ import logging
 from collections import OrderedDict
 import numpy as np
 
-import cvxpy as cp
-
 from ams.core.param import RParam
 from ams.core.service import ZonalSum, VarSelect, NumOp, NumOpDual
 from ams.routines.dcopf import DCOPF
@@ -14,140 +12,6 @@ from ams.routines.dcopf import DCOPF
 from ams.opt import Var, Constraint
 
 logger = logging.getLogger(__name__)
-
-
-# --- RTED e_fn callables (Phase 4.4) ---
-
-
-def _rted_rbu(r):
-    return r.gs @ cp.multiply(r.ug, r.pru) - r.dud == 0
-
-
-def _rted_rbd(r):
-    return r.gs @ cp.multiply(r.ug, r.prd) - r.ddd == 0
-
-
-def _rted_rru(r):
-    return cp.multiply(r.ug, r.pg + r.pru) - cp.multiply(r.ug, r.pmaxe) <= 0
-
-
-def _rted_rrd(r):
-    return cp.multiply(r.ug, -r.pg + r.prd) + cp.multiply(r.ug, r.pmine) <= 0
-
-
-def _rted_rgu(r):
-    return cp.multiply(r.ug, r.pg - r.pg0 - r.R10) <= 0
-
-
-def _rted_rgd(r):
-    return cp.multiply(r.ug, -r.pg + r.pg0 - r.R10) <= 0
-
-
-def _rted_obj(r):
-    return (r.t ** 2 * cp.sum(cp.multiply(r.c2, r.pg ** 2))
-            + cp.sum(cp.multiply(r.ug, r.c0))
-            + r.t * cp.sum(r.c1 @ r.pg + r.cru @ r.pru + r.crd @ r.prd))
-
-
-def _esd1_obj_extra(r):
-    """ESD1Base extra objective term (registered via Objective.add_term)."""
-    return r.t * cp.sum(-cp.multiply(r.cesdc, r.pce) + cp.multiply(r.cesdd, r.pde))
-
-
-def _rtedvis_obj_extra(r):
-    """RTEDVIS extra objective term."""
-    return r.t * cp.sum(cp.multiply(r.cm, r.M) + cp.multiply(r.cd, r.D))
-
-
-# DGBase
-def _dgb_cdgb(r):
-    return r.idg @ r.pg - r.pgdg == 0
-
-
-# ESD1PBase
-def _esd1p_cesd(r):
-    return r.ies @ r.pg + r.pce - r.pde == 0
-
-
-def _esd1p_SOClb(r):
-    return -r.SOC + r.SOCmin <= 0
-
-
-def _esd1p_SOCub(r):
-    return r.SOC - r.SOCmax <= 0
-
-
-def _esd1p_SOCb(r):
-    return (cp.multiply(r.En, (r.SOC - r.SOCinit))
-            - r.t * cp.multiply(r.EtaC, r.pce)
-            + r.t * cp.multiply(r.REtaD, r.pde)) == 0
-
-
-def _esd1p_SOCr(r):
-    return r.SOCend - r.SOC <= 0
-
-
-# RTEDESP
-def _rtedesp_zce(r):
-    return cp.multiply(1 - r.ucd, r.pce) <= 0
-
-
-def _rtedesp_zde(r):
-    return cp.multiply(1 - r.udd, r.pde) <= 0
-
-
-# ESD1Base (single-period)
-def _esd1b_cdb(r):
-    return r.ucd + r.udd - 1 <= 0
-
-
-def _esd1b_zce1(r):
-    return -r.zce + r.pce <= 0
-
-
-def _esd1b_zce2(r):
-    return r.zce - r.pce - r.Mb * (1 - r.ucd) <= 0
-
-
-def _esd1b_zce3(r):
-    return r.zce - r.Mb * r.ucd <= 0
-
-
-def _esd1b_zde1(r):
-    return -r.zde + r.pde <= 0
-
-
-def _esd1b_zde2(r):
-    return r.zde - r.pde - r.Mb * (1 - r.udd) <= 0
-
-
-def _esd1b_zde3(r):
-    return r.zde - r.Mb * r.udd <= 0
-
-
-def _esd1b_tcdr(r):
-    return (r.tdc0 > 0) * (r.tdc > r.tdc0) - r.ucd <= 0
-
-
-def _esd1b_tddr(r):
-    return (r.tdd0 > 0) * (r.tdd > r.tdd0) - r.udd <= 0
-
-
-# VISBase
-def _vis_Mub(r):
-    return r.M - r.Mmax <= 0
-
-
-def _vis_Dub(r):
-    return r.D - r.Dmax <= 0
-
-
-def _vis_Mreq(r):
-    return -r.gvsg @ r.M + r.dvm == 0
-
-
-def _vis_Dreq(r):
-    return -r.gvsg @ r.D + r.dvd == 0
 
 
 class SFRBase:
@@ -290,19 +154,25 @@ class RTED(SFRBase, RTEDBase, DCOPF):
         })
         # nothing to do with to map
 
-        # --- Model Section (Phase 4.4: e_fn form) ---
+        # --- Model Section ---
         # --- SFR ---
-        self.rbu.e_fn = _rted_rbu
-        self.rbd.e_fn = _rted_rbd
-        self.rru.e_fn = _rted_rru
-        self.rrd.e_fn = _rted_rrd
+        # RegUp/Dn reserve balance
+        self.rbu.e_str = 'gs @ mul(ug, pru) - dud'
+        self.rbd.e_str = 'gs @ mul(ug, prd) - ddd'
+        # RegUp/Dn reserve source
+        self.rru.e_str = 'mul(ug, (pg + pru)) - mul(ug, pmaxe)'
+        self.rrd.e_str = 'mul(ug, (-pg + prd)) + mul(ug, pmine)'
         # Gen ramping up/down
-        self.rgu.e_fn = _rted_rgu
-        self.rgd.e_fn = _rted_rgd
+        self.rgu.e_str = 'mul(ug, (pg-pg0-R10))'
+        self.rgd.e_str = 'mul(ug, (-pg+pg0-R10))'
 
         # --- objective ---
         self.obj.info = 'total generation and reserve cost'
-        self.obj.e_fn = _rted_obj
+        # NOTE: the product involved t should use ``dot``
+        cost = 't**2 dot sum(mul(c2, pg**2)) + sum(mul(ug, c0))'
+        _to_sum = 'c1 @ pg + cru @ pru + crd @ prd'
+        cost += f'+ t dot sum({_to_sum})'
+        self.obj.e_str = cost
 
     def dc2ac(self, kloss=1.0, **kwargs):
         exec_time = self.exec_time
@@ -389,9 +259,9 @@ class DGBase:
                              info='Index DG power from pg',
                              gamma='gammapdg',
                              no_parse=True, sparse=True,)
-        self.cdgb = Constraint(name='cdgb',
+        self.cdgb = Constraint(name='cdgb', is_eq=True,
                                info='Select DG power from pg',
-                               e_fn=_dgb_cdgb,)
+                               e_str='idg @ pg - pgdg',)
 
 
 class RTEDDG(DGBase, RTED):
@@ -485,23 +355,28 @@ class ESD1PBase:
                              name='ies', tex_name=r'I_{ESD}',
                              info='Index ESD from StaticGen',
                              no_parse=True)
-        self.cesd = Constraint(name='cesd',
+        self.cesd = Constraint(name='cesd', is_eq=True,
                                info='Select pce and pde from pg',
-                               e_fn=_esd1p_cesd,)
+                               e_str='ies @ pg + pce - pde',)
 
-        self.SOClb = Constraint(name='SOClb',
-                                info='SOC lower bound', e_fn=_esd1p_SOClb,)
-        self.SOCub = Constraint(name='SOCub',
-                                info='SOC upper bound', e_fn=_esd1p_SOCub,)
+        self.SOClb = Constraint(name='SOClb', is_eq=False,
+                                info='SOC lower bound',
+                                e_str='-SOC + SOCmin',)
+        self.SOCub = Constraint(name='SOCub', is_eq=False,
+                                info='SOC upper bound',
+                                e_str='SOC - SOCmax',)
 
-        self.SOCb = Constraint(name='SOCb',
-                               info='ESD1 SOC balance', e_fn=_esd1p_SOCb,)
+        SOCb = 'mul(En, (SOC - SOCinit)) - t dot mul(EtaC, pce)'
+        SOCb += '+ t dot mul(REtaD, pde)'
+        self.SOCb = Constraint(name='SOCb', is_eq=True,
+                               info='ESD1 SOC balance',
+                               e_str=SOCb,)
 
-        self.SOCr = Constraint(name='SOCr',
+        self.SOCr = Constraint(name='SOCr', is_eq=False,
                                info='ESD1 final SOC requirement',
-                               e_fn=_esd1p_SOCr,)
+                               e_str='SOCend - SOC',)
 
-        self.obj.add_term(_esd1_obj_extra)
+        self.obj.e_str += '+ t dot sum(- mul(cesdc, pce) + mul(cesdd, pde))'
 
     def _data_check(self):
         """
@@ -562,9 +437,11 @@ class RTEDESP(ESD1PBase, RTEDDG):
                           tex_name=r'u_{d,ESD}',
                           model='ESD1', no_parse=True,)
 
-        self.zce = Constraint(name='zce', info='zce bound', e_fn=_rtedesp_zce,)
+        self.zce = Constraint(name='zce', is_eq=False, info='zce bound',
+                              e_str='mul(1-ucd, pce)',)
 
-        self.zde = Constraint(name='zde', info='zde bound', e_fn=_rtedesp_zde,)
+        self.zde = Constraint(name='zde', is_eq=False, info='zde bound',
+                              e_str='mul(1-udd, pde)',)
 
     def _preinit(self):
         """
@@ -637,26 +514,35 @@ class ESD1Base(DGBase, ESD1PBase):
         self.zde.info += ':math:`z_{d,ESD}=u_{d,ESD}*p_{d,ESD}`'
 
         # --- constraints ---
-        self.cdb = Constraint(name='cdb', info='Charging decision bound',
-                              e_fn=_esd1b_cdb,)
+        self.cdb = Constraint(name='cdb', is_eq=True,
+                              info='Charging decision bound',
+                              e_str='ucd + udd - 1',)
 
-        self.zce1 = Constraint(name='zce1', info='zce bound 1', e_fn=_esd1b_zce1,)
-        self.zce2 = Constraint(name='zce2', info='zce bound 2', e_fn=_esd1b_zce2,)
-        self.zce3 = Constraint(name='zce3', info='zce bound 3', e_fn=_esd1b_zce3,)
+        self.zce1 = Constraint(name='zce1', is_eq=False, info='zce bound 1',
+                               e_str='-zce + pce',)
+        self.zce2 = Constraint(name='zce2', is_eq=False, info='zce bound 2',
+                               e_str='zce - pce - Mb dot (1-ucd)',)
+        self.zce3 = Constraint(name='zce3', is_eq=False, info='zce bound 3',
+                               e_str='zce - Mb dot ucd',)
 
-        self.zde1 = Constraint(name='zde1', info='zde bound 1', e_fn=_esd1b_zde1,)
-        self.zde2 = Constraint(name='zde2', info='zde bound 2', e_fn=_esd1b_zde2,)
-        self.zde3 = Constraint(name='zde3', info='zde bound 3', e_fn=_esd1b_zde3,)
+        self.zde1 = Constraint(name='zde1', is_eq=False, info='zde bound 1',
+                               e_str='-zde + pde',)
+        self.zde2 = Constraint(name='zde2', is_eq=False, info='zde bound 2',
+                               e_str='zde - pde - Mb dot (1-udd)',)
+        self.zde3 = Constraint(name='zde3', is_eq=False, info='zde bound 3',
+                               e_str='zde - Mb dot udd',)
 
         # force charging flag `fcd`: (tdc0 > 0) * (tdc > tdc0)
-        self.tcdr = Constraint(name='tcdr',
+        tcdr = '(tdc0 > 0) * (tdc > tdc0) - ucd'
+        self.tcdr = Constraint(name='tcdr', is_eq=False,
                                info='Minimum charging duration',
-                               e_fn=_esd1b_tcdr,)
+                               e_str=tcdr,)
 
         # force discharging flag `fdd`: (tdd0 > 0) * (tdd > tdd0)
-        self.tddr = Constraint(name='tddr',
+        tddr = '(tdd0 > 0) * (tdd > tdd0) - udd'
+        self.tddr = Constraint(name='tddr', is_eq=False,
                                info='Minimum discharging duration',
-                               e_fn=_esd1b_tddr,)
+                               e_str=tddr,)
 
 
 class RTEDES(ESD1Base, RTED):
@@ -732,14 +618,18 @@ class VISBase:
                              name='gvsg', tex_name=r'S_{g}',
                              info='Sum VSG vars vector in shape of area',
                              no_parse=True)
-        self.Mub = Constraint(name='Mub', info='M upper bound', e_fn=_vis_Mub,)
-        self.Dub = Constraint(name='Dub', info='D upper bound', e_fn=_vis_Dub,)
-        self.Mreq = Constraint(name='Mreq',
+        self.Mub = Constraint(name='Mub', is_eq=False,
+                              info='M upper bound',
+                              e_str='M - Mmax',)
+        self.Dub = Constraint(name='Dub', is_eq=False,
+                              info='D upper bound',
+                              e_str='D - Dmax',)
+        self.Mreq = Constraint(name='Mreq', is_eq=True,
                                info='Emulated inertia requirement',
-                               e_fn=_vis_Mreq,)
-        self.Dreq = Constraint(name='Dreq',
+                               e_str='-gvsg@M + dvm',)
+        self.Dreq = Constraint(name='Dreq', is_eq=True,
                                info='Emulated damping requirement',
-                               e_fn=_vis_Dreq,)
+                               e_str='-gvsg@D + dvd',)
 
         # NOTE: revise the objective function to include virtual inertia cost
 
@@ -763,7 +653,8 @@ class RTEDVIS(VISBase, RTED):
 
         # --- objective ---
         self.obj.info = 'total generation and reserve cost'
-        self.obj.add_term(_rtedvis_obj_extra)
+        vsgcost = '+ t dot sum(cm * M + cd * D)'
+        self.obj.e_str += vsgcost
 
         self.map2.update({
             'M': ('RenGen', 'M'),
